refactor(jest,rstest): move IsFunction and CalleeChainName into test_framework
