+++ internal/plugins/jest/utils/jest.go
@@ -146,16 +146,6 @@ func JestFnMemberEntriesRange(entries []ParsedJestFnMemberEntry) (core.TextRange
 	return testFramework.MemberEntriesRange(entries)
 }
 
-func getPropertyName(node *ast.Node) string {
-	switch node.Kind {
-	case ast.KindIdentifier:
-		return node.AsIdentifier().Text
-	case ast.KindPrivateIdentifier:
-		return node.AsPrivateIdentifier().Text
-	}
-	return ""
-}
-
 func GetJestKind(name string) JestFnType {
 	switch name {
 	case "describe", "fdescribe", "xdescribe":
@@ -181,77 +171,6 @@ func JestHookOrderIndex(name string) int {
 
 func GetJestFnMemberEntries(node *ast.Node) []ParsedJestFnMemberEntry {
 	return testFramework.GetMemberEntries(node)
-}
-
-// CalleeChainName returns a dotted name for a call callee expression, mirroring
-// eslint-plugin-jest getNodeName for CallExpression callees (used by expect-expect
-// assertFunctionNames matching).
-//
-// It differs from GetJestFnMemberEntries: bracket notation
-// contributes a segment only when the index matches eslint-plugin-jest's
-// supported accessor names (identifier, string literal, or no-substitution
-// template). Unsupported keys break the chain entirely. NewExpression is peeled
-// so e.g. new (require('x')).y becomes a chain.
-func CalleeChainName(expr *ast.Node) string {
-	expr = ast.SkipParentheses(expr)
-	if expr == nil {
-		return ""
-	}
-
-	switch expr.Kind {
-	case ast.KindIdentifier:
-		return expr.AsIdentifier().Text
-	case ast.KindPropertyAccessExpression:
-		pa := expr.AsPropertyAccessExpression()
-		left := CalleeChainName(pa.Expression)
-		prop := pa.Name()
-		if prop == nil {
-			return left
-		}
-		pn := getPropertyName(prop)
-		if left == "" || pn == "" {
-			return left
-		}
-		return left + "." + pn
-	case ast.KindElementAccessExpression:
-		ea := expr.AsElementAccessExpression()
-		left := CalleeChainName(ea.Expression)
-		key := calleeChainLiteralElementKey(ast.SkipParentheses(ea.ArgumentExpression))
-		if left == "" || key == "" {
-			return ""
-		}
-		return left + "." + key
-	case ast.KindCallExpression:
-		return CalleeChainName(expr.AsCallExpression().Expression)
-	case ast.KindNewExpression:
-		ne := expr.AsNewExpression()
-		if ne == nil {
-			return ""
-		}
-		return CalleeChainName(ne.Expression)
-	case ast.KindTaggedTemplateExpression:
-		return CalleeChainName(expr.AsTaggedTemplateExpression().Tag)
-	default:
-		return ""
-	}
-}
-
-// calleeChainLiteralElementKey matches eslint-plugin-jest segments for
-// MemberExpression computed with a supported accessor name only.
-func calleeChainLiteralElementKey(n *ast.Node) string {
-	if n == nil {
-		return ""
-	}
-	switch n.Kind {
-	case ast.KindIdentifier:
-		return n.AsIdentifier().Text
-	case ast.KindStringLiteral:
-		return n.AsStringLiteral().Text
-	case ast.KindNoSubstitutionTemplateLiteral:
-		return n.AsNoSubstitutionTemplateLiteral().Text
-	default:
-		return ""
-	}
 }
 
 // DefaultJestVersion is used when the Jest version cannot be resolved from settings or package.json.
@@ -476,15 +395,6 @@ func IsBooleanLiteral(node *ast.Node) (value bool, ok bool) {
 	default:
 		return false, false
 	}
-}
-
-func IsFunction(node *ast.Node) bool {
-	return ast.IsFunctionDeclaration(node) ||
-		ast.IsFunctionExpressionOrArrowFunction(node) ||
-		node.Kind == ast.KindMethodDeclaration ||
-		node.Kind == ast.KindConstructor ||
-		node.Kind == ast.KindGetAccessor ||
-		node.Kind == ast.KindSetAccessor
 }
 
 func IsMemberAccessNode(node *ast.Node) bool {
