+++ internal/plugins/rstest/utils/rstest_ast_test.go
@@ -34,100 +34,6 @@ func findFirstNodeOfKind(t *testing.T, code string, kind ast.Kind) *ast.Node {
 	return found
 }
 
-func TestIsFunction(t *testing.T) {
-	functionForms := []struct {
-		name string
-		code string
-		kind ast.Kind
-	}{
-		{"function declaration", `function f() {}`, ast.KindFunctionDeclaration},
-		{"function expression", `const f = function () {};`, ast.KindFunctionExpression},
-		{"arrow function", `const f = () => {};`, ast.KindArrowFunction},
-		{"method declaration", `class C { m() {} }`, ast.KindMethodDeclaration},
-		{"constructor", `class C { constructor() {} }`, ast.KindConstructor},
-		{"get accessor", `class C { get p() { return 1; } }`, ast.KindGetAccessor},
-		{"set accessor", `class C { set p(v) {} }`, ast.KindSetAccessor},
-	}
-	for _, form := range functionForms {
-		t.Run(form.name, func(t *testing.T) {
-			node := findFirstNodeOfKind(t, form.code, form.kind)
-			if !rstestUtils.IsFunction(node) {
-				t.Errorf("IsFunction should be true for %s", form.name)
-			}
-		})
-	}
-
-	nonFunctionForms := []struct {
-		name string
-		code string
-		kind ast.Kind
-	}{
-		{"call expression", `f();`, ast.KindCallExpression},
-		{"class declaration", `class C {}`, ast.KindClassDeclaration},
-		{"identifier", `x;`, ast.KindIdentifier},
-		{"property access", `a.b;`, ast.KindPropertyAccessExpression},
-	}
-	for _, form := range nonFunctionForms {
-		t.Run(form.name, func(t *testing.T) {
-			node := findFirstNodeOfKind(t, form.code, form.kind)
-			if rstestUtils.IsFunction(node) {
-				t.Errorf("IsFunction should be false for %s", form.name)
-			}
-		})
-	}
-
-	if rstestUtils.IsFunction(nil) {
-		t.Error("IsFunction(nil) should be false")
-	}
-}
-
-func TestCalleeChainName(t *testing.T) {
-	tests := []struct {
-		code string
-		want string
-	}{
-		{`foo();`, "foo"},
-		{`foo.bar();`, "foo.bar"},
-		{`foo.bar.baz();`, "foo.bar.baz"},
-		{`foo["bar"]();`, "foo.bar"},
-		{"foo[`bar`]();", "foo.bar"},
-		// Identifier keys are supported accessor names upstream, matching
-		// getNodeName's treatment of computed identifier properties.
-		{`foo[bar]();`, "foo.bar"},
-		// Unsupported computed keys break the chain entirely.
-		{`foo[bar.baz]();`, ""},
-		{`foo[1]();`, ""},
-		{"foo[`bar${x}`]();", ""},
-		// Calls and news are peeled.
-		{`foo().bar();`, "foo.bar"},
-		{`new Foo().bar();`, "Foo.bar"},
-		// new (require("x")).y() parses entirely as a NewExpression, so the
-		// outer call has to be forced with parentheses to exercise the peel.
-		{`(new (require("x")).y)();`, "require.y"},
-		// Parentheses and optional chaining.
-		{`(foo.bar)();`, "foo.bar"},
-		{`foo?.bar();`, "foo.bar"},
-		// Tagged templates resolve through their tag.
-		{"foo.bar`x`();", "foo.bar"},
-		// A non-identifier root yields no name, and property access on it
-		// keeps the empty left side.
-		{`"str".includes();`, ""},
-	}
-	for _, test := range tests {
-		t.Run(test.code, func(t *testing.T) {
-			call := findFirstNodeOfKind(t, test.code, ast.KindCallExpression)
-			got := rstestUtils.CalleeChainName(call.AsCallExpression().Expression)
-			if got != test.want {
-				t.Errorf("CalleeChainName = %q, want %q", got, test.want)
-			}
-		})
-	}
-
-	if got := rstestUtils.CalleeChainName(nil); got != "" {
-		t.Errorf("CalleeChainName(nil) = %q, want empty", got)
-	}
-}
-
 func TestIsPromiseChainCall(t *testing.T) {
 	tests := []struct {
 		code string
